backend/requirements.txt
flask
flask-cors
google-cloud-storage
google-cloud-aiplatform